split expr type and null info to be expr-schemable
